FIX: Correct getOwner import
As it's currently implemented, the fallback is essential here. `this` in a static method is `undefined`, and therefore cannot have an owner.

Unfortunately this plugin outlet does not seem to be tested in the plugin's own test suite. Core's test suite picked up this issue.

Followup to 3a0b46da7aa48e27e51c74e90b42c2083d89b4d5